Extract player plugin
## This Commit

Extracts some player related logic into a player plugin.

## Why?

To keep the player-centric things separate from the rest of the game
logic. Just to keep files smaller and help people scan things faster.
